+++ test/upgrade/steps.go
@@ -64,6 +64,7 @@ func (se *suiteExecution) runContinualTests(t *testing.T, num int, stopCh <-chan
 				if se.failed {
 					return
 				}
+				// The rest of this test will run in parallel with "UpgradeDowngrade" test.
 				t.Parallel()
 				handle := operation.Handler()
 				// Blocking operation.
@@ -79,85 +80,7 @@ func (se *suiteExecution) runContinualTests(t *testing.T, num int, stopCh <-chan
 	} else {
 		l.Infof("%d) 🔄 No continual tests registered. Skipping.", num)
 	}
-	//
-	//se.configuration.T.Run("ContinualTests", func(t *testing.T) {
-	//	l, err := se.configuration.logger(t)
-	//	if err != nil {
-	//		t.Fatal(err)
-	//	}
-	//	if numOps > 0 {
-	//		l.Infof(groupTemplate, num, numOps)
-	//		for i := range operations {
-	//			operation := operations[i]
-	//			l.Infof(elementTemplate, num, i+1, operation.Name())
-	//			if se.failed {
-	//				l.Debugf(skippingOperationTemplate, operation.Name())
-	//				return
-	//			}
-	//			operation.Setup()(Context{T: t}) // Note: Log is not passed here, we'll use t.Logf directly
-	//
-	//			logger, buffer := newInMemoryLoggerBuffer(se.configuration)
-	//			t.Run("Setup"+operation.Name(), func(t *testing.T) {
-	//				l, err = se.configuration.logger(t)
-	//				if err != nil {
-	//					t.Fatal(err)
-	//				}
-	//				setup(Context{T: t, Log: logger.Sugar()})
-	//			})
-	//
-	//			handler := operation.Handler()
-	//			go func() {
-	//				handler(BackgroundContext{
-	//					Log:       logger.Sugar(),
-	//					Stop:      operation.stop,
-	//					logBuffer: buffer,
-	//				})
-	//			}()
-	//			se.failed = se.failed || t.Failed()
-	//			if se.failed {
-	//				// need to dump logs here, because verify will not be executed.
-	//				l.Error(wrapLog(buffer.Dump()))
-	//				return
-	//			}
-	//		}
-	//
-	//	} else {
-	//		l.Infof("%d) 🔄 No continual tests registered. Skipping.", num)
-	//	}
-	//})
 }
-
-//func (se *suiteExecution) verifyContinualTests(num int) {
-//	testsCount := len(se.suite.tests.continual)
-//	if testsCount > 0 {
-//		se.configuration.T.Run("VerifyContinualTests", func(t *testing.T) {
-//			l, err := se.configuration.logger(t)
-//			if err != nil {
-//				t.Fatal(err)
-//			}
-//			l.Infof("%d) ✋ Verifying %d running continual tests.", num, testsCount)
-//			for i, operation := range se.suite.tests.continual {
-//				t.Run(operation.Name(), func(t *testing.T) {
-//					l, err = se.configuration.logger(t)
-//					if err != nil {
-//						t.Fatal(err)
-//					}
-//					l.Infof(`%d.%d) Verifying "%s".`, num, i+1, operation.Name())
-//					finished := make(chan struct{})
-//					operation.stop <- StopEvent{
-//						T:        t,
-//						Finished: finished,
-//						logger:   l,
-//						name:     "Stop of " + operation.Name(),
-//					}
-//					<-finished
-//					se.failed = se.failed || t.Failed()
-//					l.Debugf(`Finished "%s"`, operation.Name())
-//				})
-//			}
-//		})
-//	}
-//}
 
 func (se *suiteExecution) upgradeWith(t *testing.T, num int) {
 	se.processOperationGroup(t, operationGroup{
